+++ internal/edgeos/content.go
@@ -3,11 +3,8 @@ package edgeos
 import (
 	"bufio"
 	"bytes"
-	"errors"
 	"fmt"
 	"io"
-	"os"
-	"strings"
 	"sync"
 	"sync/atomic"
 
@@ -20,9 +17,9 @@ type IFace int
 // IFace types for labeling interface types
 const (
 	Invalid IFace = iota + 100
-	ExRtObj
 	ExDmObj
 	ExHtObj
+	ExRtObj
 	FileObj
 	PreDObj
 	PreHObj
@@ -45,42 +42,47 @@ type Contenter interface {
 	String() string
 }
 
-// ExcDomnObjects implements GetList for domain exclusions
+// ExcDomnObjects struct of *Objects for domain exclusions
 type ExcDomnObjects struct {
 	*Objects
 }
 
-// ExcHostObjects implements GetList for host exclusions
+// ExcHostObjects struct of *Objects for host exclusions
 type ExcHostObjects struct {
 	*Objects
 }
 
-// ExcRootObjects implements GetList for global domain exclusions
+// ExcRootObjects struct of *Objects for global domain exclusions
 type ExcRootObjects struct {
 	*Objects
 }
 
-// FIODataObjects implements GetList for files
+// FIODataObjects struct of *Objects for files
 type FIODataObjects struct {
 	*Objects
 }
 
-// PreDomnObjects implements GetList for pre-configured domains content
+// PreDomnObjects struct of *Objects for pre-configured domains content
 type PreDomnObjects struct {
 	*Objects
 }
 
-// PreHostObjects implements GetList for pre-configured hosts content
+// PreHostObjects struct of *Objects for pre-configured hosts content
 type PreHostObjects struct {
 	*Objects
 }
 
-// URLDomnObjects implements GetList for URLs
+// PreRootObjects struct of *Objects for pre-configured root content
+type PreRootObjects struct {
+	*Objects
+}
+
+// URLDomnObjects struct of *Objects for domain URLs
 type URLDomnObjects struct {
 	*Objects
 }
 
-// URLHostObjects implements GetList for URLs
+// URLHostObjects struct of *Objects for host URLs
 type URLHostObjects struct {
 	*Objects
 }
@@ -146,6 +148,16 @@ func (p *PreHostObjects) Find(elem string) int {
 }
 
 // Find returns the int position of an Objects' element
+func (p *PreRootObjects) Find(elem string) int {
+	for i, o := range p.src {
+		if o.name == elem {
+			return i
+		}
+	}
+	return -1
+}
+
+// Find returns the int position of an Objects' element
 func (u *URLHostObjects) Find(elem string) int {
 	for i, o := range u.src {
 		if o.name == elem {
@@ -168,8 +180,7 @@ func (u *URLDomnObjects) Find(elem string) int {
 // GetList implements the Contenter interface for ExcDomnObjects
 func (e *ExcDomnObjects) GetList() *Objects {
 	for _, o := range e.src {
-		switch o.nType {
-		case excDomn:
+		if o.nType == excDomn {
 			if o.exc != nil {
 				o.r = o.excludes()
 				o.Parms = e.Objects.Parms
@@ -182,8 +193,7 @@ func (e *ExcDomnObjects) GetList() *Objects {
 // GetList implements the Contenter interface for ExcHostObjects
 func (e *ExcHostObjects) GetList() *Objects {
 	for _, o := range e.src {
-		switch o.nType {
-		case excHost:
+		if o.nType == excHost {
 			if o.exc != nil {
 				o.r = o.excludes()
 				o.Parms = e.Objects.Parms
@@ -196,8 +206,7 @@ func (e *ExcHostObjects) GetList() *Objects {
 // GetList implements the Contenter interface for ExcRootObjects
 func (e *ExcRootObjects) GetList() *Objects {
 	for _, o := range e.src {
-		switch o.nType {
-		case excRoot:
+		if o.nType == excRoot {
 			if o.exc != nil {
 				o.r = o.excludes()
 				o.Parms = e.Objects.Parms
@@ -293,30 +302,6 @@ func (u *URLHostObjects) GetList() *Objects {
 	return u.Objects
 }
 
-// GetTotalStats displays aggregate statistics for processed sources
-func (c *Config) GetTotalStats() (dropped, extracted, kept int32) {
-	var keys []string
-
-	for k := range c.ctr {
-		keys = append(keys, k)
-	}
-
-	for _, k := range keys {
-		if c.ctr[k].kept+c.ctr[k].dropped != 0 {
-			dropped += c.ctr[k].dropped
-			extracted += c.ctr[k].extracted
-			kept += c.ctr[k].kept
-		}
-	}
-
-	if kept+dropped != 0 {
-		c.Log.Noticef("Total entries found: %d", extracted)
-		c.Log.Noticef("Total entries extracted %d", kept)
-		c.Log.Noticef("Total entries dropped %d", dropped)
-	}
-	return dropped, extracted, kept
-}
-
 // Len returns how many sources there are
 func (e *ExcDomnObjects) Len() int { return len(e.Objects.src) }
 
@@ -334,6 +319,9 @@ func (p *PreDomnObjects) Len() int { return len(p.Objects.src) }
 
 // Len returns how many sources there are
 func (p *PreHostObjects) Len() int { return len(p.Objects.src) }
+
+// Len returns how many sources there are
+func (p *PreRootObjects) Len() int { return len(p.Objects.src) }
 
 // Len returns how many sources there are
 func (u *URLDomnObjects) Len() int { return len(u.Objects.src) }
@@ -383,23 +371,20 @@ func (o *source) process() *bList {
 			continue
 		}
 	}
+	switch o.nType {
+	case domn, excDomn, excRoot:
+		o.Dex = mergeList(o.Dex, add)
+	}
 
 	// Let's do some accounting
 	atomic.AddInt32(&o.ctr[area].dropped, int32(drop))
 	atomic.AddInt32(&o.ctr[area].extracted, int32(extracted))
 	atomic.AddInt32(&o.ctr[area].kept, int32(kept))
 
-	o.Log.Infof("%s: downloaded: %d", o.name, extracted)
-	o.Log.Infof("%s: extracted: %d", o.name, kept)
-	o.Log.Infof("%s: dropped: %d", o.name, drop)
-
-	if kept == 0 {
-		o.Log.Warningf("Zero records extracted for %s, dnsmasq conf file won't be written", o.name)
-	}
-
-	switch o.nType {
-	case domn, excDomn, excRoot:
-		o.Dex = mergeList(o.Dex, add)
+	if kept != 0 && area != rootNode {
+		o.Log.Infof("%s: downloaded: %d", o.name, extracted)
+		o.Log.Infof("%s: extracted: %d", o.name, kept)
+		o.Log.Infof("%s: dropped: %d", o.name, drop)
 	}
 
 	return &bList{
@@ -411,67 +396,12 @@ func (o *source) process() *bList {
 
 func (o *source) filename(area string) string {
 	switch o.nType {
-	case excDomn, excRoot, preDomn:
+	case excDomn, excRoot, preDomn, preRoot:
 		return fmt.Sprintf(o.FnFmt, o.Dir, domains, o.name, o.Ext)
 	case excHost, preHost:
 		return fmt.Sprintf(o.FnFmt, o.Dir, hosts, o.name, o.Ext)
 	}
 	return fmt.Sprintf(o.FnFmt, o.Dir, area, o.name, o.Ext)
-}
-
-// ProcessContent processes the Contents array
-func (c *Config) ProcessContent(cts ...Contenter) error {
-	var (
-		errs      []string
-		getErrors chan error
-	)
-
-	if len(cts) < 1 {
-		return errors.New("Empty Contenter interface{} passed to ProcessContent()")
-	}
-
-	for _, ct := range cts {
-		var (
-			a, b  int32
-			area  string
-			tally = &stats{dropped: a, kept: b}
-		)
-
-		for _, o := range ct.GetList().src {
-			getErrors = make(chan error)
-
-			if o.err != nil {
-				errs = append(errs, o.err.Error())
-			}
-
-			go func(o *source) {
-				area = typeInt(o.nType)
-				c.ctr[area] = tally
-				getErrors <- o.process().writeFile()
-			}(o)
-
-			for range cts {
-				if err := <-getErrors; err != nil {
-					errs = append(errs, err.Error())
-				}
-				close(getErrors)
-			}
-		}
-
-		if area != "" {
-			if c.ctr[area].kept+c.ctr[area].dropped != 0 {
-				c.Log.Noticef("Total %s found: %d", area, c.ctr[area].extracted)
-				c.Log.Noticef("Total %s extracted %d", area, c.ctr[area].kept)
-				c.Log.Noticef("Total %s dropped %d", area, c.ctr[area].dropped)
-			}
-		}
-	}
-
-	if errs != nil {
-		return fmt.Errorf(strings.Join(errs, "\n"))
-	}
-
-	return nil
 }
 
 // SetURL sets the Object's url field value
@@ -529,6 +459,15 @@ func (p *PreHostObjects) SetURL(name, url string) {
 }
 
 // SetURL sets the Object's url field value
+func (p *PreRootObjects) SetURL(name, url string) {
+	for _, o := range p.src {
+		if o.name == name {
+			o.url = url
+		}
+	}
+}
+
+// SetURL sets the Object's url field value
 func (u *URLDomnObjects) SetURL(name, url string) {
 	for _, o := range u.src {
 		if o.name == name {
@@ -552,6 +491,7 @@ func (e *ExcRootObjects) String() string { return e.Objects.String() }
 func (f *FIODataObjects) String() string { return f.Objects.String() }
 func (p *PreDomnObjects) String() string { return p.Objects.String() }
 func (p *PreHostObjects) String() string { return p.Objects.String() }
+func (p *PreRootObjects) String() string { return p.Objects.String() }
 func (u *URLDomnObjects) String() string { return u.Objects.String() }
 func (u *URLHostObjects) String() string { return u.Objects.String() }
 
@@ -573,20 +513,4 @@ func (i IFace) String() string {
 		return urls
 	}
 	return notknown
-}
-
-// writeFile saves hosts/domains data to disk
-func (b *bList) writeFile() error {
-	if b.size == 0 {
-		return nil
-	}
-
-	w, err := os.Create(b.file)
-	if err != nil {
-		return err
-	}
-
-	defer w.Close()
-	_, err = io.Copy(w, b.r)
-	return err
 }