+++ internal/edgeos/object.go
@@ -2,29 +2,8 @@ package edgeos
 
 import (
 	"fmt"
-	"io"
 	"sort"
-	"strings"
 )
-
-// source struct for normalizing EdgeOS data.
-type source struct {
-	*Parms
-	desc     string
-	disabled bool
-	err      error
-	exc      []string
-	file     string
-	inc      []string
-	ip       string
-	ltype    string
-	name     string
-	nType    ntype
-	Objects
-	prefix string
-	r      io.Reader
-	url    string
-}
 
 // Objects is a struct of []*source
 type Objects struct {
@@ -35,30 +14,6 @@ type Objects struct {
 func (o *Objects) addObj(c *Config, node string) {
 	o.src = append(o.src, c.addInc(node))
 	o.src = append(o.src, c.tree.validate(node).src...)
-}
-
-func (o *source) area() string {
-	switch getType(o.nType).(string) {
-	case domains, PreDomns:
-		return domains
-	}
-	return hosts
-}
-
-// excludes returns an io.Reader of blacklist includes
-func (o *source) excludes() io.Reader {
-	sort.Strings(o.exc)
-	return strings.NewReader(strings.Join(o.exc, "\n"))
-}
-
-func (o *source) setFilePrefix(format string) string {
-	switch o.nType {
-	case excDomn, excRoot, preDomn:
-		return fmt.Sprintf(format, domains, o.name)
-	case excHost, preHost:
-		return fmt.Sprintf(format, hosts, o.name)
-	}
-	return fmt.Sprintf(format, getType(o.nType), o.name)
 }
 
 // Files returns a list of dnsmasq conf files from all srcs
@@ -124,12 +79,6 @@ func getLtypeDesc(l string) string {
 	}
 }
 
-// includes returns an io.Reader of blacklist includes
-func (o *source) includes() io.Reader {
-	sort.Strings(o.inc)
-	return strings.NewReader(strings.Join(o.inc, "\n"))
-}
-
 func (o *Objects) objects(c *Config, node string, ltypes ...string) {
 	var (
 		newDomns = true
@@ -170,29 +119,6 @@ func (o *Objects) Names() (s sort.StringSlice) {
 		s = append(s, obj.name)
 	}
 	sort.Sort(s)
-	return s
-}
-
-func newSource() *source {
-	return &source{
-		Objects: Objects{},
-		exc:     []string{},
-		inc:     []string{},
-	}
-}
-
-// Stringer for Object
-func (o *source) String() string {
-	s := fmt.Sprintf("\nDesc:\t %q\n", o.desc)
-	s += fmt.Sprintf("Disabled: %v\n", o.disabled)
-	s += fmt.Sprintf("File:\t %q\n", o.file)
-	s += fmt.Sprintf("IP:\t %q\n", o.ip)
-	s += fmt.Sprintf("Ltype:\t %q\n", o.ltype)
-	s += fmt.Sprintf("Name:\t %q\n", o.name)
-	s += fmt.Sprintf("nType:\t %q\n", o.nType)
-	s += fmt.Sprintf("Prefix:\t %q\n", o.prefix)
-	s += fmt.Sprintf("Type:\t %q\n", getType(o.nType))
-	s += fmt.Sprintf("URL:\t %q\n", o.url)
 	return s
 }