+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Releases
 
+### Release v1.1.6 (April 13, 2018)
+
+* Suppress log messages for predefined includes (blacklists) and excludes (whitelists)
+* Changed Debian pre-remove and post-install script algorithm to detect an existing installation and reinstall a default configuration if none exists
+
 ### Release v1.1.5 (April 10, 2018)
 
 * Add build architecture and OS information to "-version" argument